SpringBoot初始化模板v2.2.1 开发RAG向量模块&开发模型管理模块

+++ src/main/java/top/sharehome/springbootinittemplate/config/ai/spring/service/transcription/manager/TranscriptionManager.java
@@ -60,7 +60,7 @@ public class TranscriptionManager {
         return new OpenAiAudioTranscriptionModel(openAiAudioApi, OpenAiAudioTranscriptionOptions.builder()
                 .language(Objects.isNull(entity.getLanguage()) ? null : entity.getLanguage().getValue())
                 .model(entity.getModel())
-                .temperature(entity.getTemperature())
+                .temperature(entity.getTemperature().floatValue())
                 .responseFormat(entity.getFormat())
                 .build(), RetryUtils.DEFAULT_RETRY_TEMPLATE);
     }
@@ -79,7 +79,7 @@ public class TranscriptionManager {
         return new AzureOpenAiAudioTranscriptionModel(openAiClient, AzureOpenAiAudioTranscriptionOptions.builder()
                 .language(Objects.isNull(entity.getLanguage()) ? null : entity.getLanguage().getValue())
                 .deploymentName(entity.getModel())
-                .temperature(entity.getTemperature())
+                .temperature(entity.getTemperature().floatValue())
                 .responseFormat(entity.getFormat())
                 .build());
     }
